move endpoint to auditable path
